+++ src/helpers/dataGetter.js
@@ -38,9 +38,12 @@ const deleteTask = taskId => axios.delete(`${firebaseUrl}/tasks/${taskId}.json`)
 
 const addNewTask = taskObject => axios.post(`${firebaseUrl}/tasks.json`, JSON.stringify(taskObject));
 
+const updateTask = (taskObject, taskId) => axios.put(`${firebaseUrl}/tasks/${taskId}.json`, JSON.stringify(taskObject));
+
 export default {
   getAllTasksFromDb,
   getSingleTask,
   deleteTask,
   addNewTask,
+  updateTask,
 };
