+++ src/index.html
@@ -10,6 +10,8 @@
   <div id="navbar"></div>
   <div id="auth" class="text-center"></div>
   <div id="add-edit-task" class="text-center"></div>
+  <div id="show-task-form" class="btn btn-success">Show Add Task Form</div>
+  <div id="single-container" class="text-center"></div>
   <div id="lists" class="text-center">Tasks!</div>
 
 </body>
