+++ src/index.js
@@ -1,11 +1,11 @@
-// import $ from 'jquery';
+import $ from 'jquery';
 import firebase from 'firebase/app';
 import apiKeys from '../db/apiKeys.json';
 import createNavBar from './components/Navbar/navbar';
 import loginButton from './components/Auth/auth';
 import checkLoginStatus from './helpers/authHelpers';
 import initializeListsPage from './components/ListsPage/listsPage';
-import showAddForm from './components/AddEditFriends/editAddFriends';
+import showAddForm from './components/AddEditTasks/editAddTask';
 
 import 'bootstrap';
 import './index.scss';
@@ -17,6 +17,6 @@ const initializeApp = () => {
   loginButton();
   checkLoginStatus();
   initializeListsPage();
-  showAddForm();
+  $('#show-task-form').on('click', showAddForm);
 };
 initializeApp();
